clean up repo
