+++ README.md
@@ -1,15 +1,14 @@
 # swap_comparator
 
 ## Setup
-<pre>pip install -r requirements.txt</pre>
-Create `./data/` folder in the root.
+<pre>pip install -r requirements.lock</pre>
 Edit the `./example.env` into `./.env` and fill the different fields.
 
 ## Run
 <pre>python -m swap_comparator</pre>
 
 ## Test
-<pre>python -m pytest</pre>
+<pre>pytest</pre>
 
 ## Return value
 A `./data/timestamp.csv` is created / completed after a few minutes.
