+++ example.env
@@ -1,12 +1,12 @@
-RPSOdos =
+RPSOdos = '1'
 
-ID1inch =
-RPS1inch =
+ID1inch = '********************************'
+RPS1inch = '1'
 
-ID0xSwap =
-RPS0xSwap =
+ID0xSwap = '************************************'
+RPS0xSwap = '1'
 
-RPSLiFi =
-FromAddressLiFi =
+RPSLiFi = '1'
+FromAddressLiFi = '0x****************************************************************'
 
-RPSParaSwap =
+RPSParaSwap = '1'
