+++ pyproject.toml
@@ -1,0 +1,46 @@
+[project]
+name = "swap_comparator"
+version = "1.0.0"
+description = "A Python package to retreive data from several APIs."
+authors = [
+    { name = "Tijoxa", email = "92037181+Tijoxa@users.noreply.github.com" },
+]
+dependencies = ["aiohttp", "python-dotenv", "requests"]
+readme = "README.md"
+requires-python = ">=3.11"
+
+[project.optional-dependencies]
+dev = ["pytest", "pytest-env"]
+
+[tool.pytest.ini_options]
+addopts = [
+    "-rEfX",
+    "--tb=native",
+    "--capture=sys",
+    "-p",
+    "no:warnings",
+    "-p",
+    "no:pytest-shard",
+    "--assert=plain",
+]
+
+[tool.pytest.ini_options.env]
+RPSOdos = "1"
+
+ID1inch = "1"
+RPS1inch = "1"
+
+ID0xSwap = "1"
+RPS0xSwap = "1"
+
+RPSLiFi = "1"
+FromAddressLiFi = "1"
+
+RPSParaSwap = "1"
+
+
+[tool.ruff]
+line-length = 149
+
+[tool.ruff.lint]
+ignore = ["F401"]
