--- pytest.ini
@@ -1,10 +1,0 @@
-[pytest]
-addopts =
-    -rEfX
-    --tb=native
-    --capture=sys
-    -p no:warnings
-    -p no:pytest-shard
-    --assert=plain
-testpaths =
-    tests
