--- requirement.txt
@@ -1,2 +1,0 @@
-aiohttp==3.9.3
-python-dotenv==1.0.1
