+++ requirements-dev.lock
@@ -1,0 +1,38 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile pyproject.toml -o requirements-dev.lock --extra dev
+aiohttp==3.9.5
+aiosignal==1.3.1
+    # via aiohttp
+attrs==23.2.0
+    # via aiohttp
+certifi==2024.7.4
+    # via requests
+charset-normalizer==3.3.2
+    # via requests
+frozenlist==1.4.1
+    # via
+    #   aiohttp
+    #   aiosignal
+idna==3.7
+    # via
+    #   requests
+    #   yarl
+iniconfig==2.0.0
+    # via pytest
+multidict==6.0.5
+    # via
+    #   aiohttp
+    #   yarl
+packaging==24.1
+    # via pytest
+pluggy==1.5.0
+    # via pytest
+pytest==8.2.2
+    # via pytest-env
+pytest-env==1.1.3
+python-dotenv==1.0.1
+requests==2.32.3
+urllib3==2.2.2
+    # via requests
+yarl==1.9.4
+    # via aiohttp
