+++ swap_comparator/utils/__init__.py
@@ -1,0 +1,5 @@
+from swap_comparator.utils.utils_1inch import run_1inch
+from swap_comparator.utils.utils_odos import run_odos
+from swap_comparator.utils.utils_0x import run_0x
+from swap_comparator.utils.utils_lifi import run_lifi
+from swap_comparator.utils.utils_paraswap import run_paraswap
